pyproject.toml
[project]
name = "taxi-license-autocheck"
version = "0.1.0"
description = ""
authors = [
    {name = "Denis Klyachin",email = "yithling@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "2captcha-python (>=1.5.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "taxi_license_autocheck", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
